Moved LocationSelectionDialog to the AddEdit folder
