+++ app/src/main/java/com/example/vybe/AddEdit/AddEditVibeEventActivity.java
@@ -16,7 +16,6 @@ import androidx.appcompat.widget.Toolbar;
 import androidx.fragment.app.DialogFragment;
 
 import com.bumptech.glide.Glide;
-import com.example.vybe.LocationSelectionDialog;
 import com.example.vybe.MapFragment;
 import com.example.vybe.Models.vibefactory.Vibe;
 import com.example.vybe.R;
